% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selex_fxns.R
\name{doubleNorm24}
\alias{doubleNorm24}
\title{Calculate values for double normal selectivity}
\usage{
doubleNorm24(x, a, b, c, d, e, f, use_e_999, use_f_999)
}
\arguments{
\item{x}{A vector of lengths or ages}

\item{a}{The peak}

\item{b}{The top}

\item{c}{The ascending width}

\item{d}{The Descending width}

\item{e}{The initial value}

\item{f}{The final value}

\item{use_e_999}{Is -999 used for the initial value?}

\item{use_f_999}{Is -999 used for the final value?}
}
\value{
The double normal selectivity values given the parameters. This is a
numeric vector of the same length as \code{x}.
}
\description{
Calculate values for double normal selectivity
}
