% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selex_fxns.R
\name{logistic1}
\alias{logistic1}
\title{Calculate values for length logistic selectivity}
\usage{
logistic1(len, a, b)
}
\arguments{
\item{len}{A vector of lengths or ages}

\item{a}{The inflection point}

\item{b}{The 95\% width}
}
\value{
The selectivity curve as a vector
}
\description{
Calculate values for length logistic selectivity
}
