% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selex_fxns.R
\name{nonparasize6}
\alias{nonparasize6}
\title{Nonparametric size selectivity, pattern 6}
\usage{
nonparasize6(x, len_pt_first, len_pt_last, sel_val_pts, scaling_offset = 0)
}
\arguments{
\item{x}{A vector of lengths}

\item{len_pt_first}{The size for the first waypoint}

\item{len_pt_last}{The size for the last waypoint}

\item{sel_val_pts}{A numeric vector containing the log scale selectivity values at each waypoint}

\item{scaling_offset}{Should be left as 0 except for pattern 43. Defaults to 0.
each waypoint. Entered as logistic.}
}
\value{
The double normal selectivity values given the parameters. This is a
numeric vector of the same length as \code{x}.
}
\description{
Note this works for size only and not age. Note that option 43 is similar and
could be added onto this function, using 1 additional parameter
}
